Add local.sh for local Docker support

+++ scripts/refresh_project.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
+if [ -n "$LOCAL" ]; then
+  echo "LOCAL is set, not refreshing project"
+  exit 0
+fi
+
 current_branch=$(git rev-parse --abbrev-ref HEAD)
 
 git reset --hard origin/$current_branch
